requirements.txt
flask==2.3.3
pyserial==3.5
werkzeug==2.3.7 